requirements.txt
# Core Flask dependencies
flask==3.1.2
werkzeug==3.1.5

# Environment variables
python-dotenv==1.0.1

# Data validation and models
pydantic==2.12.5

# Image processing
pillow==12.1.0

# HTTP requests
requests==2.32.5

# AI/ML APIs
openai==1.58.1

# Testing dependencies (dev)
pytest==9.0.2
hypothesis==2.1.2

# Additional Flask dependencies
jinja2==3.1.6
markupsafe==3.0.3
itsdangerous==2.2.0
click==8.3.1
blinker==1.9.0